finish v1.0

+++ backend/src/dockerAPI/dockerStack.js
@@ -2,17 +2,19 @@ const util = require('util');
 const yaml = require('js-yaml');
 const fs = require('fs-extra');
 const exec = util.promisify(require('child_process').exec);
-const uuidv1 = require('uuid/v1');
+const shortid = require('shortid');
 const options2str = require('../common/utils/options2str');
 
 async function deploy(options, STACK) {
     const optionsStr = options2str(options);
     const command = `docker stack deploy ${optionsStr} ${STACK}`;
-    return await exec(command);
+    console.log(command)
+    const {stdout, stderr} = await exec(command);
+    return {stdout, stderr};
 }
-async function deployWrap(yamlData, STACK = uuidv1()) {
+async function deployWrap(yamlData, STACK = shortid.generate()) {
     const yamlRaw = yaml.safeDump(yamlData);
-    const file = `/var/tmp/${uuidv1()}.yml`
+    const file = `/var/tmp/${STACK}.yml`
     fs.outputFileSync(file, yamlRaw);
     try {
         return await deploy({'-c' : file}, STACK)
@@ -25,8 +27,13 @@ async function deployWrap(yamlData, STACK = uuidv1()) {
     }
 }
 async function rm(STACK) {
+    if(!STACK) {
+        return;
+    }
     const command = `docker stack rm ${STACK}`;
-    return await exec(command);
+    console.log(command);
+    const {stdout, stderr} = await exec(command);
+    return {stdout, stderr};
 }
 
 module.exports = {
